Add WorkQueue and Worker
