+++ pkg/crmanager/worker.go
@@ -1,0 +1,35 @@
+/*-
+* Copyright (c) 2016-2019, F5 Networks, Inc.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*    http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+ */
+
+package crmanager
+
+import log "github.com/F5Networks/k8s-bigip-ctlr/pkg/vlogger"
+
+func (crMgr *CRManager) customResourceWorker() {
+	for crMgr.processResource() {
+	}
+}
+
+func (crMgr *CRManager) processResource() bool {
+	key, quit := crMgr.rscQueue.Get()
+	if quit {
+		return false
+	}
+	defer crMgr.rscQueue.Done(key)
+	rKey := key.(rqKey)
+	log.Debugf("Processing Key: %v", rKey)
+	return true
+}
